backend/requirements.txt
Gemini
python-dotenv
websockets
openai